refac: move service implementations to dedicated package
- Moved MinioServiceImpl and UserServiceImpl from 'service' to
'service/impl' package.
- Updated imports and package declarations in test files accordingly.
